Change the DocmentActivity to MarkorBaseActivity
